refactor(web): extract LiveArtifactViewer into the file-viewer slice
Moves the fully self-contained LiveArtifactViewer component tree (the
live-artifact preview/code/data/refresh-history viewer, its present
and zoom menus, and the refresh-history/code sub-panels) out of
FileViewer.tsx into apps/web/src/features/file-viewer/, following the
ADR 0002 vertical-slice pattern (Cluster A of EXTRACTION-PLAN.md):

- Pure logic (tab list, metadata/provenance/refresh payload builders,
  refresh-event bookkeeping, status/phase describers, error mapping,
  the preview URL builder) moved to rules.ts; a slice-local
  LiveArtifactRefreshFailure error type added to types.ts so the pure
  refreshErrorMessage rule never needs to import the provider's error
  class.
- New LiveArtifactPort (ports.ts) bound in dependencies.ts to the
  existing /api/live-artifacts transport, translating the provider's
  thrown error into the in-slice failure type.
- New feature-local hooks (useLiveArtifactViewer, useLiveArtifactCode)
  own all state/effects; dumb view components render the JSX.
- Two new small provider bridges (chrome-actions-host, window-open)
  and a shared viewport-cache module extracted from FileViewer.tsx's
  module scope, since both LiveArtifactViewer and the still-inline
  HtmlViewer read from the same preview-viewport preference cache.
- FileViewer.tsx now re-exports LiveArtifactViewer and
  LiveArtifactRefreshHistoryPanel from the slice barrel so existing
  consumers (FileWorkspace.tsx, FileViewer.test.tsx) are unaffected.

FileViewer.tsx: 8773 -> 7667 lines. All 534 existing + new tests pass;
pnpm guard's vertical-slice boundary check passes.
