+++ apps/web/src/components/FileViewer.tsx
@@ -41,9 +41,6 @@ import {
 import { useT, useI18n } from '../i18n';
 import type { Dict, Locale } from '../i18n/types';
 import {
-  fetchLiveArtifact,
-  fetchLiveArtifactCode,
-  fetchLiveArtifactRefreshes,
   checkDeploymentLink,
   CLOUDFLARE_PAGES_PROVIDER_ID,
   createSocialSharePayload,
@@ -56,11 +53,8 @@ import {
   fetchProjectFileText,
   fetchProjectFileTextPreview,
   uploadProjectFiles,
-  liveArtifactPreviewUrl,
   projectFileUrl,
   projectRawUrl,
-  LiveArtifactRefreshError,
-  refreshLiveArtifact,
   updateDeployConfig,
   type WebDeployConfigResponse,
   type WebCloudflarePagesDeploySelection,
@@ -111,14 +105,7 @@ import {
   type UrlLoadDecision,
 } from './file-viewer-render-mode';
 import { resolvePoweredPreviewUrl } from '../runtime/powered-preview';
-import type {
-  LiveArtifactEventItem,
-  LiveArtifact,
-  LiveArtifactRefreshLogEntry,
-  LiveArtifactViewerTab,
-  LiveArtifactWorkspaceEntry,
-  ProjectFile,
-} from '../types';
+import type { ProjectFile } from '../types';
 import { Icon } from './Icon';
 import { RemixIcon } from './RemixIcon';
 import { SocialShareGrid } from './SocialShareGrid';
@@ -270,6 +257,9 @@ import {
   useWiredPreviewCanvasSize,
   fileVersionPreviewOptions,
   FileVersionManagerModal,
+  getCachedPreviewViewport,
+  setCachedPreviewViewport,
+  previewViewportStateKey,
 } from '../features/file-viewer';
 import type {
   InspectOverrideMap,
@@ -303,6 +293,8 @@ export {
   CommentTargetOverlay,
   CommentSidePanel,
   fileVersionPreviewOptions,
+  LiveArtifactViewer,
+  LiveArtifactRefreshHistoryPanel,
 } from '../features/file-viewer';
 export type { InspectOverrideEntry, InspectOverrideMap, ManualEditPendingStyleSave } from '../features/file-viewer';
 
@@ -355,13 +347,11 @@ const HTML_ROUTING_TEXT_PREVIEW_LIMIT = 96 * 1024;
 // getComputedStyle on the iframe side.
 const MAX_CACHED_SLIDE_STATES = 64;
 const htmlPreviewSlideState = new Map<string, SlideState>();
-const MAX_CACHED_PREVIEW_VIEWPORTS = 128;
 // Grace window before the inspect hover card is torn down. Long enough to absorb
 // the async iframe mouseout (od:comment-leave) that fires when the pointer slides
 // onto the card or hops back onto the element under it, short enough to read as
 // an immediate dismiss when the pointer really leaves.
 const HOVER_CARD_DISMISS_DELAY_MS = 80;
-const htmlPreviewViewportState = new Map<string, PreviewViewportId>();
 const MARKDOWN_COPY_BLOCK_ATTR = 'data-copy-code-block';
 const MARKDOWN_COPY_BUTTON_CLASS = 'markdown-code-copy';
 const MARKDOWN_COPY_TOAST_CLASS = 'markdown-code-toast';
@@ -559,18 +549,6 @@ async function requestPreviewSnapshotWithRetry(iframe: HTMLIFrameElement): Promi
   return null;
 }
 
-function previewViewportStateKey(projectId: string, file: Pick<ProjectFile, 'name' | 'path'>): string {
-  return `${projectId}:${file.path || file.name}`;
-}
-
-function setPreviewViewportCached(key: string, viewport: PreviewViewportId) {
-  htmlPreviewViewportState.set(key, viewport);
-  if (htmlPreviewViewportState.size > MAX_CACHED_PREVIEW_VIEWPORTS) {
-    const oldest = htmlPreviewViewportState.keys().next().value;
-    if (oldest != null) htmlPreviewViewportState.delete(oldest);
-  }
-}
-
 interface Props {
   projectId: string;
   projectKind: TrackingProjectKind;
@@ -723,719 +701,6 @@ export function FileViewer({
   return <BinaryViewer projectId={projectId} file={file} />;
 }
 
-export function LiveArtifactViewer({
-  projectId,
-  liveArtifact,
-  liveArtifactEvents = [],
-  onRefreshArtifacts,
-}: {
-  projectId: string;
-  liveArtifact: LiveArtifactWorkspaceEntry;
-  liveArtifactEvents?: LiveArtifactEventItem[];
-  onRefreshArtifacts?: () => Promise<void> | void;
-}) {
-  const t = useT();
-  const tabs = useMemo(() => liveArtifactViewerTabs(t), [t]);
-  const [mode, setMode] = useState<LiveArtifactViewerTab>('preview');
-  const [detail, setDetail] = useState<LiveArtifact | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [reloadKey, setReloadKey] = useState(0);
-  const [zoom, setZoom] = useState(100);
-  const liveArtifactViewportKey = `${projectId}:live-artifact:${liveArtifact.artifactId}`;
-  const [previewViewport, setPreviewViewportState] = useState<PreviewViewportId>(
-    () => htmlPreviewViewportState.get(liveArtifactViewportKey) ?? 'desktop',
-  );
-  const setPreviewViewport = useCallback((viewport: PreviewViewportId) => {
-    setPreviewViewportCached(liveArtifactViewportKey, viewport);
-    setPreviewViewportState(viewport);
-  }, [liveArtifactViewportKey]);
-  const [previewBodyRef, previewBodySize] = useWiredPreviewCanvasSize<HTMLDivElement>();
-  const iframeRef = useRef<HTMLIFrameElement | null>(null);
-  const [refreshing, setRefreshing] = useState(false);
-  const [refreshError, setRefreshError] = useState<string | null>(null);
-  const [refreshSuccess, setRefreshSuccess] = useState<string | null>(null);
-  const [refreshEvents, setRefreshEvents] = useState<LiveArtifactRefreshEvent[]>([]);
-  const [refreshHistory, setRefreshHistory] = useState<LiveArtifactRefreshLogEntry[]>([]);
-  const [presentMenuOpen, setPresentMenuOpen] = useState(false);
-  const [zoomMenuOpen, setZoomMenuOpen] = useState(false);
-  const zoomMenuRef = useRef<HTMLDivElement | null>(null);
-  const [inTabPresent, setInTabPresent] = useState(false);
-  const presentWrapRef = useRef<HTMLDivElement | null>(null);
-  const [chromeActionsHost, setChromeActionsHost] = useState<HTMLElement | null>(null);
-  useEffect(() => {
-    if (typeof document === 'undefined') return;
-    setChromeActionsHost(resolveChromeActionsHost());
-  }, []);
-  useEffect(() => {
-    if (!presentMenuOpen) return;
-    const onPointer = (e: MouseEvent) => {
-      const target = e.target as HTMLElement | null;
-      if (!target) return;
-      if (target.closest('.present-wrap')) return;
-      setPresentMenuOpen(false);
-    };
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') setPresentMenuOpen(false);
-    };
-    document.addEventListener('mousedown', onPointer);
-    document.addEventListener('keydown', onKey);
-    return () => {
-      document.removeEventListener('mousedown', onPointer);
-      document.removeEventListener('keydown', onKey);
-    };
-  }, [presentMenuOpen]);
-
-  useEffect(() => {
-    setRefreshError(null);
-    setRefreshSuccess(null);
-    setRefreshEvents([]);
-  }, [projectId, liveArtifact.artifactId]);
-
-  useEffect(() => {
-    setPreviewViewportState(htmlPreviewViewportState.get(liveArtifactViewportKey) ?? 'desktop');
-  }, [liveArtifactViewportKey]);
-
-  useEffect(() => {
-    if (!refreshSuccess) return;
-    const timeout = window.setTimeout(() => setRefreshSuccess(null), 6000);
-    return () => window.clearTimeout(timeout);
-  }, [refreshSuccess]);
-
-  const processedLiveArtifactEventIdRef = useRef(0);
-
-  useEffect(() => {
-    const pendingEvents = liveArtifactEvents.filter((item) => item.id > processedLiveArtifactEventIdRef.current);
-    if (pendingEvents.length === 0) return;
-    processedLiveArtifactEventIdRef.current = pendingEvents[pendingEvents.length - 1]?.id ?? processedLiveArtifactEventIdRef.current;
-
-    for (const { event: liveArtifactEvent } of pendingEvents) {
-    if (
-      (liveArtifactEvent.kind !== 'live_artifact' && liveArtifactEvent.kind !== 'live_artifact_refresh') ||
-      liveArtifactEvent.projectId !== projectId ||
-      liveArtifactEvent.artifactId !== liveArtifact.artifactId
-    ) {
-      continue;
-    }
-
-    if (liveArtifactEvent.kind === 'live_artifact') {
-      setRefreshError(null);
-      if (liveArtifactEvent.action === 'deleted') {
-        setRefreshSuccess(`Live artifact deleted: ${liveArtifactEvent.title}`);
-        continue;
-      }
-      setRefreshSuccess(
-        liveArtifactEvent.action === 'created'
-          ? `Live artifact created: ${liveArtifactEvent.title}`
-          : `Live artifact updated: ${liveArtifactEvent.title}`,
-      );
-      void fetchLiveArtifact(projectId, liveArtifact.artifactId).then((next) => {
-        if (next) setDetail(next);
-      });
-      void fetchLiveArtifactRefreshes(projectId, liveArtifact.artifactId).then(setRefreshHistory);
-      setReloadKey((n) => n + 1);
-      continue;
-    }
-
-    if (liveArtifactEvent.phase === 'started') {
-      setRefreshing(true);
-      setRefreshError(null);
-      setRefreshSuccess(null);
-      setRefreshEvents((prev) => appendRefreshEvent(prev, { phase: 'started' }));
-      continue;
-    }
-
-    if (liveArtifactEvent.phase === 'failed') {
-      setRefreshing(false);
-      setRefreshError(liveArtifactEvent.error ?? t('liveArtifact.refresh.genericFailure'));
-      setRefreshEvents((prev) =>
-        appendRefreshEvent(prev, {
-          phase: 'failed',
-          error: liveArtifactEvent.error ?? undefined,
-        }),
-      );
-      void fetchLiveArtifact(projectId, liveArtifact.artifactId).then((next) => {
-        if (next) setDetail(next);
-      });
-      void fetchLiveArtifactRefreshes(projectId, liveArtifact.artifactId).then(setRefreshHistory);
-      continue;
-    }
-
-    setRefreshing(false);
-    setRefreshError(null);
-    setRefreshEvents((prev) =>
-      appendRefreshEvent(prev, {
-        phase: 'succeeded',
-        refreshedSourceCount: liveArtifactEvent.refreshedSourceCount ?? 0,
-      }),
-    );
-    if ((liveArtifactEvent.refreshedSourceCount ?? 0) > 0) {
-      setRefreshSuccess(t('liveArtifact.refresh.successOne'));
-    } else {
-      setRefreshError(t('liveArtifact.refresh.noSourceTitle'));
-    }
-    void fetchLiveArtifact(projectId, liveArtifact.artifactId).then((next) => {
-      if (next) setDetail(next);
-    });
-    void fetchLiveArtifactRefreshes(projectId, liveArtifact.artifactId).then(setRefreshHistory);
-    setReloadKey((n) => n + 1);
-    }
-  }, [liveArtifactEvents, liveArtifact.artifactId, projectId, t]);
-
-  useEffect(() => {
-    let cancelled = false;
-    setLoading(true);
-    setDetail(null);
-    void fetchLiveArtifact(projectId, liveArtifact.artifactId).then((next) => {
-      if (cancelled) return;
-      setDetail(next);
-      setLoading(false);
-    });
-    void fetchLiveArtifactRefreshes(projectId, liveArtifact.artifactId).then((next) => {
-      if (!cancelled) setRefreshHistory(next);
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [projectId, liveArtifact.artifactId, liveArtifact.updatedAt]);
-
-  const previewUrl = useMemo(
-    () => `${liveArtifactPreviewUrl(projectId, liveArtifact.artifactId)}&v=${reloadKey}`,
-    [projectId, liveArtifact.artifactId, reloadKey],
-  );
-  const previewScale = zoom / 100;
-
-  // Instrument the live-artifact iframe so failed loads — usually a
-  // missing artifact file or a stuck `od://` resolver — surface in
-  // PostHog. iframe load errors don't propagate to window.error, so
-  // observability/install.ts cannot catch them globally.
-  useEffect(() => {
-    if (mode !== 'preview') return undefined;
-    const node = iframeRef.current;
-    if (!node) return undefined;
-    return trackIframeLoad({
-      iframe: node,
-      surface: 'live_artifact_preview',
-      artifactId: liveArtifact.artifactId,
-      projectId,
-    });
-  }, [mode, previewUrl, liveArtifact.artifactId, projectId]);
-
-  async function handleRefresh() {
-    if (refreshing) return;
-    setRefreshing(true);
-    setRefreshError(null);
-    setRefreshSuccess(null);
-    setRefreshEvents((prev) => appendRefreshEvent(prev, { phase: 'started' }));
-    try {
-      const result = await refreshLiveArtifact(projectId, liveArtifact.artifactId);
-      setDetail(result.artifact);
-      void fetchLiveArtifactRefreshes(projectId, liveArtifact.artifactId).then(setRefreshHistory);
-      setReloadKey((n) => n + 1);
-      setRefreshEvents((prev) =>
-        appendRefreshEvent(prev, {
-          phase: 'succeeded',
-          refreshedSourceCount: result.refresh.refreshedSourceCount,
-        }),
-      );
-      if (result.refresh.refreshedSourceCount > 0) {
-        setRefreshSuccess(t('liveArtifact.refresh.successOne'));
-      } else {
-        setRefreshError(t('liveArtifact.refresh.noSourceTitle'));
-      }
-      await onRefreshArtifacts?.();
-    } catch (error) {
-      const message = refreshErrorMessage(error, t);
-      setRefreshError(message);
-      setRefreshEvents((prev) => appendRefreshEvent(prev, { phase: 'failed', error: message }));
-    } finally {
-      setRefreshing(false);
-    }
-  }
-
-  const dataPayload = detail?.document?.dataJson ?? null;
-  const currentRefreshStatus = detail?.refreshStatus ?? liveArtifact.refreshStatus;
-  const isRunning = refreshing || currentRefreshStatus === 'running';
-
-  const presentInThisTab = () => {
-    setPresentMenuOpen(false);
-    setMode('preview');
-    setInTabPresent(true);
-  };
-  const presentFullscreen = () => {
-    setPresentMenuOpen(false);
-    setMode('preview');
-    const target = previewBodyRef.current ?? iframeRef.current;
-    if (target?.requestFullscreen) {
-      void target.requestFullscreen().catch(() => {});
-    }
-  };
-  const presentNewTab = () => {
-    setPresentMenuOpen(false);
-    if (typeof window === 'undefined') return;
-    window.open(liveArtifactPreviewUrl(projectId, liveArtifact.artifactId), '_blank', 'noopener,noreferrer');
-  };
-  useEffect(() => {
-    if (!inTabPresent) return;
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') setInTabPresent(false);
-    };
-    document.addEventListener('keydown', onKey);
-    return () => document.removeEventListener('keydown', onKey);
-  }, [inTabPresent]);
-
-  useEffect(() => {
-    if (!zoomMenuOpen) return;
-    const onDocClick = (e: MouseEvent) => {
-      if (!zoomMenuRef.current) return;
-      if (!zoomMenuRef.current.contains(e.target as Node)) setZoomMenuOpen(false);
-    };
-    const onKey = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') setZoomMenuOpen(false);
-    };
-    document.addEventListener('mousedown', onDocClick);
-    document.addEventListener('keydown', onKey);
-    return () => {
-      document.removeEventListener('mousedown', onDocClick);
-      document.removeEventListener('keydown', onKey);
-    };
-  }, [zoomMenuOpen]);
-
-  return (
-    <div className={`viewer html-viewer live-artifact-viewer${inTabPresent ? ' is-tab-present' : ''}`}>
-      {((node: ReactNode) => (
-        chromeActionsHost ? createPortal(node, chromeActionsHost) : node
-      ))(
-        <div className="present-wrap chrome-present-wrap" ref={presentWrapRef}>
-          <button
-            className="chrome-action chrome-action-secondary chrome-action-icon present-trigger od-tooltip"
-            aria-haspopup="menu"
-            aria-expanded={presentMenuOpen}
-            aria-label={t('fileViewer.present')}
-            data-tooltip={t('fileViewer.present')}
-            data-tooltip-placement="bottom"
-            title={t('fileViewer.present')}
-            onClick={() => setPresentMenuOpen((v) => !v)}
-          >
-            <RemixIcon name="slideshow-3-line" size={15} />
-          </button>
-          {presentMenuOpen ? (
-            <div className="present-menu" role="menu">
-              <button role="menuitem" onClick={presentInThisTab}>
-                <span className="present-icon"><RemixIcon name="eye-line" size={14} /></span>{' '}
-                {t('fileViewer.presentInTab')}
-              </button>
-              <button role="menuitem" onClick={presentFullscreen}>
-                <span className="present-icon"><RemixIcon name="play-line" size={14} /></span>{' '}
-                {t('fileViewer.presentFullscreen')}
-              </button>
-              <button role="menuitem" onClick={presentNewTab}>
-                <span className="present-icon"><RemixIcon name="share-forward-line" size={14} /></span>{' '}
-                {t('fileViewer.presentNewTab')}
-              </button>
-            </div>
-          ) : null}
-        </div>
-      )}
-      {inTabPresent ? (
-        <button
-          type="button"
-          className="present-exit-btn"
-          onClick={() => setInTabPresent(false)}
-          title={t('common.exitFullscreen')}
-          aria-label={t('common.exitFullscreen')}
-        >
-          <Icon name="close" size={14} />
-        </button>
-      ) : null}
-      <div className="viewer-toolbar">
-        <div className="viewer-toolbar-left">
-            <button
-              type="button"
-              className="icon-only od-tooltip"
-              onClick={() => setReloadKey((n) => n + 1)}
-              title={`${t('fileViewer.reload')} ${t('fileViewer.preview')}`}
-              data-tooltip={`${t('fileViewer.reload')} ${t('fileViewer.preview')}`}
-              data-tooltip-placement="bottom"
-              aria-label={`${t('fileViewer.reloadAria')} ${t('fileViewer.preview')}`}
-            >
-            <Icon name="reload" size={14} />
-          </button>
-        </div>
-        <div className="viewer-toolbar-actions">
-          <div className="viewer-tabs">
-            {tabs.map((tab) => (
-              <button
-                key={tab.id}
-                type="button"
-                className={`viewer-tab ${mode === tab.id ? 'active' : ''}`}
-                onClick={() => setMode(tab.id)}
-              >
-                {tab.label}
-              </button>
-            ))}
-          </div>
-          <div
-            className="viewer-preview-controls"
-            data-active={mode === 'preview' ? 'true' : 'false'}
-            aria-hidden={mode === 'preview' ? undefined : true}
-          >
-            <span className="viewer-divider" aria-hidden />
-            <PreviewViewportControls
-              viewport={previewViewport}
-              onViewport={setPreviewViewport}
-              t={t}
-              tabIndex={mode === 'preview' ? 0 : -1}
-            />
-            <span className="viewer-divider" aria-hidden />
-            <div className="zoom-menu viewer-toolbar-zoom" ref={zoomMenuRef}>
-              <button
-                type="button"
-                className="viewer-action zoom-trigger od-tooltip"
-                aria-haspopup="menu"
-                aria-expanded={zoomMenuOpen}
-                title={t('fileViewer.resetZoom')}
-                data-tooltip={t('fileViewer.resetZoom')}
-                data-tooltip-placement="bottom"
-                tabIndex={mode === 'preview' ? 0 : -1}
-                onClick={() => setZoomMenuOpen((v) => !v)}
-              >
-                <span style={{ fontVariantNumeric: 'tabular-nums' }}>{zoom}%</span>
-              </button>
-              {zoomMenuOpen && mode === 'preview' ? (
-                <div className="zoom-menu-popover" role="menu">
-                  {[50, 75, 100, 125, 150, 200].map((level) => (
-                    <button
-                      key={level}
-                      type="button"
-                      className={`zoom-menu-item${zoom === level ? ' active' : ''}`}
-                      role="menuitem"
-                      onClick={() => {
-                        setZoom(level);
-                        setZoomMenuOpen(false);
-                      }}
-                    >
-                      <span style={{ fontVariantNumeric: 'tabular-nums' }}>{level}%</span>
-                      {zoom === level ? (
-                        <Icon name="check" size={13} />
-                      ) : null}
-                    </button>
-                  ))}
-                </div>
-              ) : null}
-            </div>
-            <span className="viewer-divider" aria-hidden />
-            <a
-              className="ghost-link"
-              href={liveArtifactPreviewUrl(projectId, liveArtifact.artifactId)}
-              target="_blank"
-              rel="noreferrer noopener"
-              tabIndex={mode === 'preview' ? 0 : -1}
-            >
-              {t('fileViewer.open')}
-            </a>
-          </div>
-          <span className="viewer-divider" aria-hidden />
-          <button
-            type="button"
-            className="viewer-action primary"
-            data-running={isRunning ? 'true' : 'false'}
-            onClick={() => void handleRefresh()}
-            disabled={isRunning}
-            aria-busy={isRunning}
-            aria-label={isRunning ? t('liveArtifact.refresh.running') : t('liveArtifact.refresh.button')}
-            title={
-              isRunning
-                ? t('liveArtifact.refresh.running')
-                : t('liveArtifact.refresh.buttonTitle')
-            }
-          >
-            <Icon name={isRunning ? 'spinner' : 'reload'} size={13} />
-            <span>{isRunning ? t('liveArtifact.refresh.running') : t('liveArtifact.refresh.button')}</span>
-          </button>
-        </div>
-      </div>
-      <div className="viewer-body" ref={previewBodyRef}>
-        {refreshError ? (
-          <LiveArtifactRefreshNotice
-            tone="error"
-            message={refreshError}
-            action={t('liveArtifact.refresh.failureAction')}
-          />
-        ) : refreshSuccess ? (
-          <LiveArtifactRefreshNotice
-            tone="success"
-            message={refreshSuccess}
-            action={t('liveArtifact.refresh.successAction')}
-            onDismiss={() => setRefreshSuccess(null)}
-            dismissLabel={t('common.close')}
-          />
-        ) : isRunning ? (
-          <LiveArtifactRefreshNotice
-            tone="running"
-            message={t('liveArtifact.refresh.runningMessage')}
-            action={t('liveArtifact.refresh.runningAction')}
-          />
-        ) : currentRefreshStatus === 'failed' ? (
-          <LiveArtifactRefreshNotice
-            tone="error"
-            message={t('liveArtifact.refresh.previousFailure', { message: t('liveArtifact.refresh.genericFailure') })}
-            action={t('liveArtifact.refresh.failureAction')}
-          />
-        ) : null}
-        <div
-          className={`live-artifact-preview-layer preview-viewport preview-viewport-${previewViewport}`}
-          data-active={mode === 'preview' ? 'true' : 'false'}
-          aria-hidden={mode === 'preview' ? undefined : true}
-          style={previewViewportStyle(previewViewport, previewScale, previewBodySize)}
-        >
-          <div className="preview-frame-clip">
-            <div style={previewScaleShellStyle(previewViewport, previewScale)}>
-              <PreviewDrawOverlay>
-                <iframe
-                  ref={iframeRef}
-                  data-testid="live-artifact-preview-frame"
-                  title={liveArtifact.title}
-                  sandbox="allow-scripts allow-popups allow-downloads"
-                  src={previewUrl}
-                />
-              </PreviewDrawOverlay>
-            </div>
-          </div>
-        </div>
-        {mode !== 'preview' && loading ? (
-          <div className="viewer-empty">{t('fileViewer.loading')}</div>
-        ) : mode === 'code' ? (
-          <LiveArtifactCodePanel
-            projectId={projectId}
-            artifactId={liveArtifact.artifactId}
-            reloadKey={reloadKey}
-          />
-        ) : mode === 'data' ? (
-          <JsonPanel value={dataPayload} emptyLabel={t('liveArtifact.viewer.dataEmpty')} />
-        ) : (
-          <LiveArtifactRefreshHistoryPanel
-            liveArtifact={detail}
-            fallbackRefreshStatus={liveArtifact.refreshStatus}
-            fallbackLastRefreshedAt={liveArtifact.lastRefreshedAt}
-            isRunning={isRunning}
-            sessionEvents={refreshEvents}
-            persistedEvents={refreshHistory}
-          />
-        )}
-      </div>
-    </div>
-  );
-}
-
-function LiveArtifactRefreshNotice({
-  tone,
-  message,
-  action,
-  onDismiss,
-  dismissLabel,
-}: {
-  tone: 'running' | 'success' | 'error';
-  message: string;
-  action: string;
-  onDismiss?: () => void;
-  dismissLabel?: string;
-}) {
-  return (
-    <div
-      className={`live-artifact-refresh-notice ${tone}`}
-      role={tone === 'error' ? 'alert' : 'status'}
-      aria-label={`${message} ${action}`}
-    >
-      <span className="live-artifact-refresh-notice-copy">
-        <strong>{message}</strong>
-        <span>{action}</span>
-      </span>
-      {onDismiss ? (
-        <button type="button" className="icon-only" onClick={onDismiss} aria-label={dismissLabel}>
-          ×
-        </button>
-      ) : null}
-    </div>
-  );
-}
-
-function refreshErrorMessage(error: unknown, t: TranslateFn): string {
-  if (error instanceof LiveArtifactRefreshError && error.status === 0) {
-    return t('liveArtifact.refresh.networkFailure');
-  }
-  if (error instanceof LiveArtifactRefreshError && error.code === 'LIVE_ARTIFACT_REFRESH_UNAVAILABLE') {
-    return t('liveArtifact.refresh.noSourceTitle');
-  }
-  if (error instanceof Error && error.message.length > 0) return error.message;
-  return t('liveArtifact.refresh.genericFailure');
-}
-
-function liveArtifactViewerTabs(t: TranslateFn): Array<{ id: LiveArtifactViewerTab; label: string }> {
-  return [
-    { id: 'preview', label: t('liveArtifact.viewer.tabPreview') },
-    { id: 'code', label: t('liveArtifact.viewer.tabCode') },
-    { id: 'data', label: t('liveArtifact.viewer.tabData') },
-    { id: 'refresh-history', label: t('liveArtifact.viewer.tabRefreshHistory') },
-  ];
-}
-
-type LiveArtifactCodeVariant = 'template' | 'rendered-source';
-
-function LiveArtifactCodePanel({
-  projectId,
-  artifactId,
-  reloadKey,
-}: {
-  projectId: string;
-  artifactId: string;
-  reloadKey: number;
-}) {
-  const t = useT();
-  const [variant, setVariant] = useState<LiveArtifactCodeVariant>('template');
-  const [code, setCode] = useState<string | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [failed, setFailed] = useState(false);
-
-  useEffect(() => {
-    let cancelled = false;
-    setLoading(true);
-    setFailed(false);
-    setCode(null);
-    void fetchLiveArtifactCode(projectId, artifactId, variant).then((next) => {
-      if (cancelled) return;
-      setCode(next);
-      setFailed(next == null);
-      setLoading(false);
-    });
-    return () => {
-      cancelled = true;
-    };
-  }, [artifactId, projectId, reloadKey, variant]);
-
-  return (
-    <div className="live-artifact-code-panel">
-      <div className="live-artifact-code-header">
-        <div className="live-artifact-code-copy">
-          <strong>
-            {variant === 'template'
-              ? t('liveArtifact.viewer.code.templateHeading')
-              : t('liveArtifact.viewer.code.renderedHeading')}
-          </strong>
-          <span>
-            {variant === 'template'
-              ? t('liveArtifact.viewer.code.templateHelp')
-              : t('liveArtifact.viewer.code.renderedHelp')}
-          </span>
-        </div>
-        <div
-          className="viewer-tabs live-artifact-code-tabs"
-          aria-label={t('liveArtifact.viewer.code.variantAria')}
-        >
-          <button
-            type="button"
-            className={`viewer-tab ${variant === 'template' ? 'active' : ''}`}
-            onClick={() => setVariant('template')}
-          >
-            {t('liveArtifact.viewer.code.variantTemplate')}
-          </button>
-          <button
-            type="button"
-            className={`viewer-tab ${variant === 'rendered-source' ? 'active' : ''}`}
-            onClick={() => setVariant('rendered-source')}
-          >
-            {t('liveArtifact.viewer.code.variantRendered')}
-          </button>
-        </div>
-      </div>
-      {loading ? (
-        <div className="viewer-empty">{t('liveArtifact.viewer.code.loading')}</div>
-      ) : failed ? (
-        <div className="viewer-empty">{t('liveArtifact.viewer.code.unavailable')}</div>
-      ) : code && code.trim().length > 0 ? (
-        <pre className="viewer-source">{code}</pre>
-      ) : (
-        <div className="viewer-empty">{t('liveArtifact.viewer.code.empty')}</div>
-      )}
-    </div>
-  );
-}
-
-function liveArtifactMetadataPayload(liveArtifact: LiveArtifact): unknown {
-  return {
-    artifact: {
-      id: liveArtifact.id,
-      title: liveArtifact.title,
-      slug: liveArtifact.slug,
-      status: liveArtifact.status,
-      pinned: liveArtifact.pinned,
-      preview: liveArtifact.preview,
-      refreshStatus: liveArtifact.refreshStatus,
-      createdAt: liveArtifact.createdAt,
-      updatedAt: liveArtifact.updatedAt,
-      lastRefreshedAt: liveArtifact.lastRefreshedAt,
-    },
-    document: liveArtifact.document
-      ? {
-          format: liveArtifact.document.format,
-          templatePath: liveArtifact.document.templatePath,
-          generatedPreviewPath: liveArtifact.document.generatedPreviewPath,
-          dataPath: liveArtifact.document.dataPath,
-          dataSchemaJson: liveArtifact.document.dataSchemaJson,
-          sourceJson: liveArtifact.document.sourceJson,
-        }
-      : null,
-  };
-}
-
-function liveArtifactProvenancePayload(liveArtifact: LiveArtifact): unknown {
-  return {
-    documentSource: liveArtifact.document?.sourceJson ?? null,
-  };
-}
-
-function liveArtifactRefreshPayload(liveArtifact: LiveArtifact): unknown {
-  return {
-    refreshStatus: liveArtifact.refreshStatus,
-    lastRefreshedAt: liveArtifact.lastRefreshedAt ?? null,
-  };
-}
-
-type LiveArtifactRefreshStatus = LiveArtifact['refreshStatus'];
-
-interface LiveArtifactRefreshEvent {
-  id: number;
-  phase: 'started' | 'succeeded' | 'failed';
-  at: number;
-  durationMs?: number;
-  refreshedSourceCount?: number;
-  error?: string;
-}
-
-let refreshEventSequence = 0;
-
-function appendRefreshEvent(
-  prev: LiveArtifactRefreshEvent[],
-  next: Omit<LiveArtifactRefreshEvent, 'id' | 'at' | 'durationMs'>,
-): LiveArtifactRefreshEvent[] {
-  const at = Date.now();
-  refreshEventSequence += 1;
-  const event: LiveArtifactRefreshEvent = { ...next, id: refreshEventSequence, at };
-  if (next.phase !== 'started') {
-    // Pair with the most recent 'started' to compute duration.
-    for (let i = prev.length - 1; i >= 0; i -= 1) {
-      const candidate = prev[i];
-      if (candidate && candidate.phase === 'started') {
-        event.durationMs = Math.max(0, at - candidate.at);
-        break;
-      }
-    }
-  }
-  // Cap at 25 entries to keep the panel lightweight.
-  const MAX = 25;
-  const combined = [...prev, event];
-  return combined.length > MAX ? combined.slice(combined.length - MAX) : combined;
-}
-
 function hasSeenExportReadyNudge(projectId: string, fileName: string): boolean {
   try {
     return window.sessionStorage.getItem(exportReadyNudgeKey(projectId, fileName)) === '1';
@@ -1451,377 +716,6 @@ function markExportReadyNudgeSeen(projectId: string, fileName: string) {
     // Ignore storage-denied contexts; the in-memory state still prevents loops.
   }
 }
-
-interface RefreshStatusDescriptor {
-  label: string;
-  tone: 'neutral' | 'running' | 'success' | 'warning' | 'error';
-  description: string;
-}
-
-function describeRefreshStatus(
-  status: LiveArtifactRefreshStatus,
-  t: TranslateFn,
-): RefreshStatusDescriptor {
-  switch (status) {
-    case 'running':
-      return {
-        label: t('liveArtifact.refresh.statusRunning'),
-        tone: 'running',
-        description: t('liveArtifact.refresh.statusRunningDescription'),
-      };
-    case 'succeeded':
-      return {
-        label: t('liveArtifact.refresh.statusSucceeded'),
-        tone: 'success',
-        description: t('liveArtifact.refresh.statusSucceededDescription'),
-      };
-    case 'failed':
-      return {
-        label: t('liveArtifact.refresh.statusFailed'),
-        tone: 'error',
-        description: t('liveArtifact.refresh.statusFailedDescription'),
-      };
-    case 'idle':
-      return {
-        label: t('liveArtifact.refresh.statusReady'),
-        tone: 'neutral',
-        description: t('liveArtifact.refresh.statusReadyDescription'),
-      };
-    case 'never':
-    default:
-      return {
-        label: t('liveArtifact.refresh.statusNever'),
-        tone: 'warning',
-        description: t('liveArtifact.refresh.statusNeverDescription'),
-      };
-  }
-}
-
-function describeEventPhase(
-  event: LiveArtifactRefreshEvent,
-  t: TranslateFn,
-): { label: string; tone: 'running' | 'success' | 'error' } {
-  if (event.phase === 'started')
-    return { label: t('liveArtifact.refresh.eventStarted'), tone: 'running' };
-  if (event.phase === 'succeeded')
-    return { label: t('liveArtifact.refresh.eventSucceeded'), tone: 'success' };
-  return { label: t('liveArtifact.refresh.eventFailed'), tone: 'error' };
-}
-
-function describePersistedStatus(
-  status: LiveArtifactRefreshLogEntry['status'],
-  t: TranslateFn,
-): string {
-  switch (status) {
-    case 'succeeded':
-      return t('liveArtifact.refresh.persistedStatusSucceeded');
-    case 'running':
-      return t('liveArtifact.refresh.persistedStatusRunning');
-    case 'failed':
-      return t('liveArtifact.refresh.persistedStatusFailed');
-    case 'cancelled':
-      return t('liveArtifact.refresh.persistedStatusCancelled');
-    case 'skipped':
-      return t('liveArtifact.refresh.persistedStatusSkipped');
-    default: {
-      const exhaustive: never = status;
-      return exhaustive;
-    }
-  }
-}
-
-export function LiveArtifactRefreshHistoryPanel({
-  liveArtifact,
-  fallbackRefreshStatus,
-  fallbackLastRefreshedAt,
-  isRunning,
-  sessionEvents,
-  persistedEvents = [],
-}: {
-  liveArtifact: LiveArtifact | null;
-  fallbackRefreshStatus: LiveArtifactRefreshStatus;
-  fallbackLastRefreshedAt?: string;
-  isRunning: boolean;
-  sessionEvents: LiveArtifactRefreshEvent[];
-  persistedEvents?: LiveArtifactRefreshLogEntry[];
-}) {
-  const t = useT();
-  const { locale } = useI18n();
-  const [now, setNow] = useState(() => Date.now());
-
-  useEffect(() => {
-    // Keep relative timestamps fresh; 30s cadence is enough for "x minutes ago" feel.
-    const id = window.setInterval(() => setNow(Date.now()), 30_000);
-    return () => window.clearInterval(id);
-  }, []);
-
-  const status: LiveArtifactRefreshStatus = isRunning
-    ? 'running'
-    : liveArtifact?.refreshStatus ?? fallbackRefreshStatus;
-  const descriptor = describeRefreshStatus(status, t);
-  const lastRefreshedAt = liveArtifact?.lastRefreshedAt ?? fallbackLastRefreshedAt;
-  const createdAt = liveArtifact?.createdAt;
-  const updatedAt = liveArtifact?.updatedAt;
-  const documentSource = liveArtifact?.document?.sourceJson ?? null;
-  const reversedEvents = [...sessionEvents].reverse();
-  const reversedPersistedEvents = [...persistedEvents].reverse().slice(0, 25);
-  const rawDebugPayload = liveArtifact
-    ? {
-        refresh: liveArtifactRefreshPayload(liveArtifact),
-        metadata: liveArtifactMetadataPayload(liveArtifact),
-        provenance: liveArtifactProvenancePayload(liveArtifact),
-      }
-    : null;
-
-  return (
-    <div className="live-artifact-refresh-panel">
-      <section className="live-artifact-refresh-hero">
-        <div className="live-artifact-refresh-hero-main">
-          <span
-            className={`live-artifact-badge refresh-status tone-${descriptor.tone}`}
-            data-testid="live-artifact-refresh-status-badge"
-          >
-            {descriptor.label}
-          </span>
-          <p className="live-artifact-refresh-hero-desc">{descriptor.description}</p>
-        </div>
-        <div className="live-artifact-refresh-hero-meta">
-          <div className="live-artifact-refresh-hero-metric">
-            <span className="live-artifact-refresh-label">
-              {t('liveArtifact.refresh.heroLastRefreshedLabel')}
-            </span>
-            {lastRefreshedAt ? (
-              <>
-                <span className="live-artifact-refresh-value">
-                  {formatRelativeTime(lastRefreshedAt, now, locale, t) ?? '—'}
-                </span>
-                <span
-                  className="live-artifact-refresh-sub"
-                  title={formatAbsoluteDateTime(lastRefreshedAt) ?? undefined}
-                >
-                  {formatAbsoluteDateTime(lastRefreshedAt) ?? ''}
-                </span>
-              </>
-            ) : (
-              <span className="live-artifact-refresh-value muted">
-                {t('liveArtifact.refresh.heroLastRefreshedNever')}
-              </span>
-            )}
-          </div>
-        </div>
-      </section>
-
-      <section className="live-artifact-refresh-facts">
-        <LiveArtifactRefreshFact
-          label={t('liveArtifact.refresh.factCreated')}
-          iso={createdAt}
-          emptyLabel={t('liveArtifact.refresh.factUnknown')}
-          now={now}
-          locale={locale}
-          t={t}
-        />
-        <LiveArtifactRefreshFact
-          label={t('liveArtifact.refresh.factLastUpdated')}
-          iso={updatedAt}
-          emptyLabel={t('liveArtifact.refresh.factUnknown')}
-          now={now}
-          locale={locale}
-          t={t}
-        />
-      </section>
-
-      <section className="live-artifact-refresh-section">
-        <header className="live-artifact-refresh-section-header">
-          <h4>{t('liveArtifact.refresh.persistedTitle')}</h4>
-          <span className="live-artifact-refresh-hint">
-            {t('liveArtifact.refresh.persistedHint')}
-          </span>
-        </header>
-        {reversedPersistedEvents.length === 0 ? (
-          <div className="live-artifact-refresh-empty">
-            {t('liveArtifact.refresh.persistedEmpty')}
-          </div>
-        ) : (
-          <ol className="live-artifact-refresh-timeline">
-            {reversedPersistedEvents.map((event) => {
-              const tone = event.status === 'succeeded'
-                ? 'success'
-                : event.status === 'running'
-                  ? 'running'
-                  : event.status === 'failed' || event.status === 'cancelled'
-                    ? 'error'
-                    : 'running';
-              const duration = formatDurationMs(event.durationMs);
-              return (
-                <li key={`${event.refreshId}:${event.sequence}`} className={`live-artifact-refresh-event tone-${tone}`}>
-                  <span className="live-artifact-refresh-event-dot" aria-hidden />
-                  <div className="live-artifact-refresh-event-body">
-                    <div className="live-artifact-refresh-event-row">
-                      <span className={`live-artifact-badge refresh-status tone-${tone}`}>
-                        {describePersistedStatus(event.status, t)}
-                      </span>
-                      <strong>{event.step}</strong>
-                      <span className="live-artifact-refresh-event-time">
-                        {formatRelativeTime(event.startedAt, now, locale, t)
-                          ?? t('liveArtifact.refresh.justNow')}
-                      </span>
-                    </div>
-                    <div className="live-artifact-refresh-event-meta">
-                      <span>{event.refreshId}</span>
-                      {duration ? <span>{duration}</span> : null}
-                      {event.error?.message ? <span>{event.error.message}</span> : null}
-                    </div>
-                  </div>
-                </li>
-              );
-            })}
-          </ol>
-        )}
-      </section>
-
-      <section className="live-artifact-refresh-section">
-        <header className="live-artifact-refresh-section-header">
-          <h4>{t('liveArtifact.refresh.sessionTitle')}</h4>
-          <span className="live-artifact-refresh-hint">
-            {t('liveArtifact.refresh.sessionHint')}
-          </span>
-        </header>
-        {reversedEvents.length === 0 ? (
-          <div className="live-artifact-refresh-empty">
-            {t('liveArtifact.refresh.timelineEmpty')}
-          </div>
-        ) : (
-          <ol className="live-artifact-refresh-timeline">
-            {reversedEvents.map((event) => {
-              const phase = describeEventPhase(event, t);
-              const duration = formatDurationMs(event.durationMs);
-              const refreshedCount = event.refreshedSourceCount ?? 0;
-              return (
-                <li key={event.id} className={`live-artifact-refresh-event tone-${phase.tone}`}>
-                  <span className="live-artifact-refresh-event-dot" aria-hidden />
-                  <div className="live-artifact-refresh-event-body">
-                    <div className="live-artifact-refresh-event-row">
-                      <span
-                        className={`live-artifact-badge refresh-status tone-${phase.tone}`}
-                      >
-                        {phase.label}
-                      </span>
-                      <span
-                        className="live-artifact-refresh-event-time"
-                        title={formatAbsoluteDateTime(event.at) ?? undefined}
-                      >
-                        {formatRelativeTime(event.at, now, locale, t) ?? ''}
-                      </span>
-                    </div>
-                    <div className="live-artifact-refresh-event-detail">
-                      {event.phase === 'succeeded' ? (
-                        <span>
-                          {t(
-                            refreshedCount === 1
-                              ? 'liveArtifact.refresh.sourcesUpdatedOne'
-                              : 'liveArtifact.refresh.sourcesUpdatedMany',
-                            { n: refreshedCount },
-                          )}
-                          {duration ? ` · ${duration}` : ''}
-                        </span>
-                      ) : event.phase === 'failed' ? (
-                        <span>
-                          {event.error ?? t('liveArtifact.refresh.genericFailure')}
-                          {duration ? ` · ${duration}` : ''}
-                        </span>
-                      ) : (
-                        <span>{t('liveArtifact.refresh.eventStartedDetail')}</span>
-                      )}
-                    </div>
-                  </div>
-                </li>
-              );
-            })}
-          </ol>
-        )}
-      </section>
-
-      {documentSource ? (
-        <section className="live-artifact-refresh-section">
-          <header className="live-artifact-refresh-section-header">
-            <h4>{t('liveArtifact.refresh.docSourceTitle')}</h4>
-            <span className="live-artifact-refresh-hint">
-              {t('liveArtifact.refresh.docSourceHint')}
-            </span>
-          </header>
-          <dl className="live-artifact-refresh-kv">
-            <div>
-              <dt>{t('liveArtifact.refresh.docSourceType')}</dt>
-              <dd>{documentSource.type}</dd>
-            </div>
-            {documentSource.toolName ? (
-              <div>
-                <dt>{t('liveArtifact.refresh.docSourceTool')}</dt>
-                <dd>
-                  <code>{documentSource.toolName}</code>
-                </dd>
-              </div>
-            ) : null}
-            {documentSource.connector ? (
-              <div>
-                <dt>{t('liveArtifact.refresh.docSourceConnector')}</dt>
-                <dd>
-                  {documentSource.connector.accountLabel ??
-                    documentSource.connector.connectorId}
-                </dd>
-              </div>
-            ) : null}
-          </dl>
-        </section>
-      ) : null}
-
-      {rawDebugPayload != null ? (
-        <details className="live-artifact-refresh-raw">
-          <summary>{t('liveArtifact.refresh.debugSummary')}</summary>
-          <p className="live-artifact-refresh-raw-note">
-            {t('liveArtifact.refresh.debugNote')}
-          </p>
-          <pre className="viewer-source">{JSON.stringify(rawDebugPayload, null, 2)}</pre>
-        </details>
-      ) : null}
-    </div>
-  );
-}
-
-function LiveArtifactRefreshFact({
-  label,
-  iso,
-  value,
-  helper,
-  emptyLabel,
-  now,
-  locale,
-  t,
-}: {
-  label: string;
-  iso?: string;
-  value?: string;
-  helper?: string;
-  emptyLabel?: string;
-  now?: number;
-  locale?: Locale;
-  t?: TranslateFn;
-}) {
-  const relative = iso !== undefined ? formatRelativeTime(iso, now, locale, t) : null;
-  const absolute = iso !== undefined ? formatAbsoluteDateTime(iso) : null;
-  const resolved = value ?? relative ?? emptyLabel ?? '—';
-  const sub = helper ?? (iso !== undefined ? absolute ?? '' : '');
-  return (
-    <div className="live-artifact-refresh-fact">
-      <span className="live-artifact-refresh-label">{label}</span>
-      <span className="live-artifact-refresh-value" title={absolute ?? undefined}>
-        {resolved}
-      </span>
-      {sub ? <span className="live-artifact-refresh-sub">{sub}</span> : null}
-    </div>
-  );
-}
-
 
 // fileVersionSourceLabel/fileVersionSourceToTracking/fileVersionPreviewOptions and
 // FileVersionManagerModal now live in the file-viewer slice (the modal's
@@ -2108,10 +1002,10 @@ function HtmlViewer({
   const [zoom, setZoom] = useState(100);
   const fileViewportKey = previewViewportStateKey(projectId, file);
   const [previewViewport, setPreviewViewportState] = useState<PreviewViewportId>(
-    () => htmlPreviewViewportState.get(fileViewportKey) ?? 'desktop',
+    () => getCachedPreviewViewport(fileViewportKey) ?? 'desktop',
   );
   const setPreviewViewport = useCallback((viewport: PreviewViewportId) => {
-    setPreviewViewportCached(fileViewportKey, viewport);
+    setCachedPreviewViewport(fileViewportKey, viewport);
     setPreviewViewportState(viewport);
   }, [fileViewportKey]);
   const [zoomMenuOpen, setZoomMenuOpen] = useState(false);
@@ -2139,7 +1033,7 @@ function HtmlViewer({
   });
 
   useEffect(() => {
-    setPreviewViewportState(htmlPreviewViewportState.get(fileViewportKey) ?? 'desktop');
+    setPreviewViewportState(getCachedPreviewViewport(fileViewportKey) ?? 'desktop');
   }, [fileViewportKey]);
   const [deployment, setDeployment] = useState<WebDeploymentInfo | null>(null);
   const [deploymentsByProvider, setDeploymentsByProvider] = useState<Partial<Record<WebDeployProviderId, WebDeploymentInfo>>>({});
