+++ apps/web/src/features/file-viewer/components/LiveArtifactCodePanelView.tsx
@@ -1,0 +1,66 @@
+// Dumb presentational view for the live-artifact code panel: props in, JSX
+// out. State/transport live in useLiveArtifactCode.hooks.ts.
+import type { LiveArtifactCodeVariant, TranslateFn } from '../types';
+
+export function LiveArtifactCodePanelView({
+  t,
+  variant,
+  onSetVariant,
+  code,
+  loading,
+  failed,
+}: {
+  t: TranslateFn;
+  variant: LiveArtifactCodeVariant;
+  onSetVariant: (variant: LiveArtifactCodeVariant) => void;
+  code: string | null;
+  loading: boolean;
+  failed: boolean;
+}) {
+  return (
+    <div className="live-artifact-code-panel">
+      <div className="live-artifact-code-header">
+        <div className="live-artifact-code-copy">
+          <strong>
+            {variant === 'template'
+              ? t('liveArtifact.viewer.code.templateHeading')
+              : t('liveArtifact.viewer.code.renderedHeading')}
+          </strong>
+          <span>
+            {variant === 'template'
+              ? t('liveArtifact.viewer.code.templateHelp')
+              : t('liveArtifact.viewer.code.renderedHelp')}
+          </span>
+        </div>
+        <div
+          className="viewer-tabs live-artifact-code-tabs"
+          aria-label={t('liveArtifact.viewer.code.variantAria')}
+        >
+          <button
+            type="button"
+            className={`viewer-tab ${variant === 'template' ? 'active' : ''}`}
+            onClick={() => onSetVariant('template')}
+          >
+            {t('liveArtifact.viewer.code.variantTemplate')}
+          </button>
+          <button
+            type="button"
+            className={`viewer-tab ${variant === 'rendered-source' ? 'active' : ''}`}
+            onClick={() => onSetVariant('rendered-source')}
+          >
+            {t('liveArtifact.viewer.code.variantRendered')}
+          </button>
+        </div>
+      </div>
+      {loading ? (
+        <div className="viewer-empty">{t('liveArtifact.viewer.code.loading')}</div>
+      ) : failed ? (
+        <div className="viewer-empty">{t('liveArtifact.viewer.code.unavailable')}</div>
+      ) : code && code.trim().length > 0 ? (
+        <pre className="viewer-source">{code}</pre>
+      ) : (
+        <div className="viewer-empty">{t('liveArtifact.viewer.code.empty')}</div>
+      )}
+    </div>
+  );
+}
