+++ apps/web/src/features/file-viewer/components/LiveArtifactRefreshHistoryPanel.tsx
@@ -1,0 +1,276 @@
+// Dumb (read-only) panel showing a live artifact's refresh status, session +
+// persisted refresh timelines, document-source facts, and a raw debug dump.
+import { useEffect, useState } from 'react';
+import { useI18n, useT } from '../../../i18n';
+import type { LiveArtifact, LiveArtifactRefreshLogEntry } from '../../../types';
+import {
+  describeEventPhase,
+  describePersistedStatus,
+  describeRefreshStatus,
+  liveArtifactMetadataPayload,
+  liveArtifactProvenancePayload,
+  liveArtifactRefreshPayload,
+} from '../rules';
+import { formatAbsoluteDateTime, formatDurationMs, formatRelativeTime } from '../formatters';
+import { LiveArtifactRefreshFact } from './LiveArtifactRefreshFact';
+import type { LiveArtifactRefreshEvent } from '../types';
+
+type LiveArtifactRefreshStatus = LiveArtifact['refreshStatus'];
+
+export function LiveArtifactRefreshHistoryPanel({
+  liveArtifact,
+  fallbackRefreshStatus,
+  fallbackLastRefreshedAt,
+  isRunning,
+  sessionEvents,
+  persistedEvents = [],
+}: {
+  liveArtifact: LiveArtifact | null;
+  fallbackRefreshStatus: LiveArtifactRefreshStatus;
+  fallbackLastRefreshedAt?: string;
+  isRunning: boolean;
+  sessionEvents: LiveArtifactRefreshEvent[];
+  persistedEvents?: LiveArtifactRefreshLogEntry[];
+}) {
+  const t = useT();
+  const { locale } = useI18n();
+  const [now, setNow] = useState(() => Date.now());
+
+  useEffect(() => {
+    // Keep relative timestamps fresh; 30s cadence is enough for "x minutes ago" feel.
+    const id = setInterval(() => setNow(Date.now()), 30_000);
+    return () => clearInterval(id);
+  }, []);
+
+  const status: LiveArtifactRefreshStatus = isRunning
+    ? 'running'
+    : liveArtifact?.refreshStatus ?? fallbackRefreshStatus;
+  const descriptor = describeRefreshStatus(status, t);
+  const lastRefreshedAt = liveArtifact?.lastRefreshedAt ?? fallbackLastRefreshedAt;
+  const createdAt = liveArtifact?.createdAt;
+  const updatedAt = liveArtifact?.updatedAt;
+  const documentSource = liveArtifact?.document?.sourceJson ?? null;
+  const reversedEvents = [...sessionEvents].reverse();
+  const reversedPersistedEvents = [...persistedEvents].reverse().slice(0, 25);
+  const rawDebugPayload = liveArtifact
+    ? {
+        refresh: liveArtifactRefreshPayload(liveArtifact),
+        metadata: liveArtifactMetadataPayload(liveArtifact),
+        provenance: liveArtifactProvenancePayload(liveArtifact),
+      }
+    : null;
+
+  return (
+    <div className="live-artifact-refresh-panel">
+      <section className="live-artifact-refresh-hero">
+        <div className="live-artifact-refresh-hero-main">
+          <span
+            className={`live-artifact-badge refresh-status tone-${descriptor.tone}`}
+            data-testid="live-artifact-refresh-status-badge"
+          >
+            {descriptor.label}
+          </span>
+          <p className="live-artifact-refresh-hero-desc">{descriptor.description}</p>
+        </div>
+        <div className="live-artifact-refresh-hero-meta">
+          <div className="live-artifact-refresh-hero-metric">
+            <span className="live-artifact-refresh-label">
+              {t('liveArtifact.refresh.heroLastRefreshedLabel')}
+            </span>
+            {lastRefreshedAt ? (
+              <>
+                <span className="live-artifact-refresh-value">
+                  {formatRelativeTime(lastRefreshedAt, now, locale, t) ?? '—'}
+                </span>
+                <span
+                  className="live-artifact-refresh-sub"
+                  title={formatAbsoluteDateTime(lastRefreshedAt) ?? undefined}
+                >
+                  {formatAbsoluteDateTime(lastRefreshedAt) ?? ''}
+                </span>
+              </>
+            ) : (
+              <span className="live-artifact-refresh-value muted">
+                {t('liveArtifact.refresh.heroLastRefreshedNever')}
+              </span>
+            )}
+          </div>
+        </div>
+      </section>
+
+      <section className="live-artifact-refresh-facts">
+        <LiveArtifactRefreshFact
+          label={t('liveArtifact.refresh.factCreated')}
+          iso={createdAt}
+          emptyLabel={t('liveArtifact.refresh.factUnknown')}
+          now={now}
+          locale={locale}
+          t={t}
+        />
+        <LiveArtifactRefreshFact
+          label={t('liveArtifact.refresh.factLastUpdated')}
+          iso={updatedAt}
+          emptyLabel={t('liveArtifact.refresh.factUnknown')}
+          now={now}
+          locale={locale}
+          t={t}
+        />
+      </section>
+
+      <section className="live-artifact-refresh-section">
+        <header className="live-artifact-refresh-section-header">
+          <h4>{t('liveArtifact.refresh.persistedTitle')}</h4>
+          <span className="live-artifact-refresh-hint">
+            {t('liveArtifact.refresh.persistedHint')}
+          </span>
+        </header>
+        {reversedPersistedEvents.length === 0 ? (
+          <div className="live-artifact-refresh-empty">
+            {t('liveArtifact.refresh.persistedEmpty')}
+          </div>
+        ) : (
+          <ol className="live-artifact-refresh-timeline">
+            {reversedPersistedEvents.map((event) => {
+              const tone = event.status === 'succeeded'
+                ? 'success'
+                : event.status === 'running'
+                  ? 'running'
+                  : event.status === 'failed' || event.status === 'cancelled'
+                    ? 'error'
+                    : 'running';
+              const duration = formatDurationMs(event.durationMs);
+              return (
+                <li key={`${event.refreshId}:${event.sequence}`} className={`live-artifact-refresh-event tone-${tone}`}>
+                  <span className="live-artifact-refresh-event-dot" aria-hidden />
+                  <div className="live-artifact-refresh-event-body">
+                    <div className="live-artifact-refresh-event-row">
+                      <span className={`live-artifact-badge refresh-status tone-${tone}`}>
+                        {describePersistedStatus(event.status, t)}
+                      </span>
+                      <strong>{event.step}</strong>
+                      <span className="live-artifact-refresh-event-time">
+                        {formatRelativeTime(event.startedAt, now, locale, t)
+                          ?? t('liveArtifact.refresh.justNow')}
+                      </span>
+                    </div>
+                    <div className="live-artifact-refresh-event-meta">
+                      <span>{event.refreshId}</span>
+                      {duration ? <span>{duration}</span> : null}
+                      {event.error?.message ? <span>{event.error.message}</span> : null}
+                    </div>
+                  </div>
+                </li>
+              );
+            })}
+          </ol>
+        )}
+      </section>
+
+      <section className="live-artifact-refresh-section">
+        <header className="live-artifact-refresh-section-header">
+          <h4>{t('liveArtifact.refresh.sessionTitle')}</h4>
+          <span className="live-artifact-refresh-hint">
+            {t('liveArtifact.refresh.sessionHint')}
+          </span>
+        </header>
+        {reversedEvents.length === 0 ? (
+          <div className="live-artifact-refresh-empty">
+            {t('liveArtifact.refresh.timelineEmpty')}
+          </div>
+        ) : (
+          <ol className="live-artifact-refresh-timeline">
+            {reversedEvents.map((event) => {
+              const phase = describeEventPhase(event, t);
+              const duration = formatDurationMs(event.durationMs);
+              const refreshedCount = event.refreshedSourceCount ?? 0;
+              return (
+                <li key={event.id} className={`live-artifact-refresh-event tone-${phase.tone}`}>
+                  <span className="live-artifact-refresh-event-dot" aria-hidden />
+                  <div className="live-artifact-refresh-event-body">
+                    <div className="live-artifact-refresh-event-row">
+                      <span
+                        className={`live-artifact-badge refresh-status tone-${phase.tone}`}
+                      >
+                        {phase.label}
+                      </span>
+                      <span
+                        className="live-artifact-refresh-event-time"
+                        title={formatAbsoluteDateTime(event.at) ?? undefined}
+                      >
+                        {formatRelativeTime(event.at, now, locale, t) ?? ''}
+                      </span>
+                    </div>
+                    <div className="live-artifact-refresh-event-detail">
+                      {event.phase === 'succeeded' ? (
+                        <span>
+                          {t(
+                            refreshedCount === 1
+                              ? 'liveArtifact.refresh.sourcesUpdatedOne'
+                              : 'liveArtifact.refresh.sourcesUpdatedMany',
+                            { n: refreshedCount },
+                          )}
+                          {duration ? ` · ${duration}` : ''}
+                        </span>
+                      ) : event.phase === 'failed' ? (
+                        <span>
+                          {event.error ?? t('liveArtifact.refresh.genericFailure')}
+                          {duration ? ` · ${duration}` : ''}
+                        </span>
+                      ) : (
+                        <span>{t('liveArtifact.refresh.eventStartedDetail')}</span>
+                      )}
+                    </div>
+                  </div>
+                </li>
+              );
+            })}
+          </ol>
+        )}
+      </section>
+
+      {documentSource ? (
+        <section className="live-artifact-refresh-section">
+          <header className="live-artifact-refresh-section-header">
+            <h4>{t('liveArtifact.refresh.docSourceTitle')}</h4>
+            <span className="live-artifact-refresh-hint">
+              {t('liveArtifact.refresh.docSourceHint')}
+            </span>
+          </header>
+          <dl className="live-artifact-refresh-kv">
+            <div>
+              <dt>{t('liveArtifact.refresh.docSourceType')}</dt>
+              <dd>{documentSource.type}</dd>
+            </div>
+            {documentSource.toolName ? (
+              <div>
+                <dt>{t('liveArtifact.refresh.docSourceTool')}</dt>
+                <dd>
+                  <code>{documentSource.toolName}</code>
+                </dd>
+              </div>
+            ) : null}
+            {documentSource.connector ? (
+              <div>
+                <dt>{t('liveArtifact.refresh.docSourceConnector')}</dt>
+                <dd>
+                  {documentSource.connector.accountLabel ??
+                    documentSource.connector.connectorId}
+                </dd>
+              </div>
+            ) : null}
+          </dl>
+        </section>
+      ) : null}
+
+      {rawDebugPayload != null ? (
+        <details className="live-artifact-refresh-raw">
+          <summary>{t('liveArtifact.refresh.debugSummary')}</summary>
+          <p className="live-artifact-refresh-raw-note">
+            {t('liveArtifact.refresh.debugNote')}
+          </p>
+          <pre className="viewer-source">{JSON.stringify(rawDebugPayload, null, 2)}</pre>
+        </details>
+      ) : null}
+    </div>
+  );
+}
