+++ apps/web/src/features/file-viewer/components/LiveArtifactRefreshNotice.tsx
@@ -1,0 +1,33 @@
+// Dumb banner shown atop the live-artifact preview for a refresh's running/
+// success/error state.
+export function LiveArtifactRefreshNotice({
+  tone,
+  message,
+  action,
+  onDismiss,
+  dismissLabel,
+}: {
+  tone: 'running' | 'success' | 'error';
+  message: string;
+  action: string;
+  onDismiss?: () => void;
+  dismissLabel?: string;
+}) {
+  return (
+    <div
+      className={`live-artifact-refresh-notice ${tone}`}
+      role={tone === 'error' ? 'alert' : 'status'}
+      aria-label={`${message} ${action}`}
+    >
+      <span className="live-artifact-refresh-notice-copy">
+        <strong>{message}</strong>
+        <span>{action}</span>
+      </span>
+      {onDismiss ? (
+        <button type="button" className="icon-only" onClick={onDismiss} aria-label={dismissLabel}>
+          ×
+        </button>
+      ) : null}
+    </div>
+  );
+}
