+++ apps/web/src/features/file-viewer/dependencies.ts
@@ -3,11 +3,16 @@
 // import `providers/` — everything else in the slice depends on a port, so
 // swapping the adapter (or a fake in tests) touches only this file.
 import {
+  fetchLiveArtifact,
+  fetchLiveArtifactCode,
+  fetchLiveArtifactRefreshes,
   fetchProjectFilePreview,
   fetchProjectFiles,
   fetchProjectFileText,
   fetchProjectFileVersion,
   fetchProjectFileVersions,
+  LiveArtifactRefreshError,
+  refreshLiveArtifact as refreshLiveArtifactTransport,
   restoreProjectFileVersion,
 } from '../../providers/registry';
 import { copyTextFileToClipboard } from '../../providers/file-viewer/clipboard';
@@ -19,19 +24,25 @@ import {
 } from '../../providers/file-viewer/outside-dismiss';
 import { observeElementSize } from '../../providers/file-viewer/element-size';
 import { documentBodyPortalRoot } from '../../providers/file-viewer/portal-root';
+import { resolveChromeActionsHost } from '../../providers/file-viewer/chrome-actions-host';
+import { openInNewTab } from '../../providers/file-viewer/window-open';
 import { saveTemplate } from '../../providers/templates';
 import { copyToClipboard } from '../../lib/copy-to-clipboard';
+import { LiveArtifactRefreshFailure } from './types';
 import type {
+  ChromeActionsHostPort,
   ClipboardPort,
   DismissPort,
   DocumentPreviewPort,
   ElementSizePort,
   FileTextPort,
   FileVersionsPort,
+  LiveArtifactPort,
   PortalPort,
   ProjectFilesPort,
   ShareLinkClipboardPort,
   TemplateSavePort,
+  WindowOpenPort,
 } from './ports';
 
 /** Default binding: the real `/api/projects/:id/files/:name/preview` transport. */
@@ -87,4 +98,37 @@ export const templateSavePort: TemplateSavePort = {
 /** Default binding: the real Clipboard API + textarea-fallback, boolean-result adapter. */
 export const shareLinkClipboardPort: ShareLinkClipboardPort = {
   copyToClipboard,
+};
+
+/**
+ * Default binding: the real `/api/live-artifacts` transport. `refreshLiveArtifact`
+ * wraps the provider's `refreshLiveArtifact`/catches its `LiveArtifactRefreshError`
+ * and rethrows the in-slice `LiveArtifactRefreshFailure` (types.ts), so callers
+ * inside the slice never need to import the provider's error class to do an
+ * `instanceof` check.
+ */
+export const liveArtifactPort: LiveArtifactPort = {
+  fetchLiveArtifact,
+  fetchLiveArtifactRefreshes,
+  fetchLiveArtifactCode,
+  async refreshLiveArtifact(projectId, artifactId) {
+    try {
+      return await refreshLiveArtifactTransport(projectId, artifactId);
+    } catch (error) {
+      if (error instanceof LiveArtifactRefreshError) {
+        throw new LiveArtifactRefreshFailure(error.message, error.status, error.code);
+      }
+      throw error;
+    }
+  },
+};
+
+/** Default binding: the real app-chrome file-actions portal-slot bridge. */
+export const chromeActionsHostPort: ChromeActionsHostPort = {
+  getChromeActionsHost: resolveChromeActionsHost,
+};
+
+/** Default binding: the real `window.open` new-tab bridge. */
+export const windowOpenPort: WindowOpenPort = {
+  openInNewTab,
 };
